+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - `machoutils` helper utility that works similar to `pefileutils`, but for Mach-O files. (@ddash-ct)
 - Add `macho` property to `FileObject`. (@ddash-ct)
 - `lief>=0.16.0;python_version>="3.9"` dependency for `macho`
+- Add metadata fields for `MessageBox`, `Telegram`, and `SmartContract` parameters. (@ddash-ct)
 
 
 ## [3.14.0] - 2024-05-29